+++ R/utils.R
@@ -21,6 +21,22 @@ mutate_when_missing <- function(.data, ...) {
   return(.data)
 }
 
+#' Expand columns
+#'
+#' Expands a data frame with columns. Added columns will have `NA_character_`
+#' values, existing columns of the same name will not be overwritten.
+#'
+#' @param df A data frame.
+#' @param colnames A character vector of column names.
+#' @return Data frame expanded with columns that were not yet present.
+#' @family helper functions
+#' @noRd
+expand_cols <- function(df, colnames) {
+  cols_to_add <- setdiff(colnames, colnames(df))
+  df[, cols_to_add] <- NA_character_
+  return(df)
+}
+
 #' Replace deploymentIDs
 #'
 #' Replaces deploymentIDs in deployments, media and observations..
@@ -507,4 +523,3 @@ remove_duplicates <- function(data_list) {
     })
 
   return(unique_data_list)
-}