+++ tests/testthat/test-example_dataset.R
@@ -1,10 +1,10 @@
-test_that("example_dataset() returns the expected dataset", {
+test_that("example_dataset() returns the latest example dataset", {
   skip_if_offline()
-  camtrapdp_1.0 <-
-    "https://raw.githubusercontent.com/tdwg/camtrap-dp/1.0/example/datapackage.json"
+  camtrapdp_latest <-
+    "https://raw.githubusercontent.com/tdwg/camtrap-dp/1.0.1/example/datapackage.json"
   expect_identical(
     example_dataset(),
-    read_camtrapdp(camtrapdp_1.0)
+    read_camtrapdp(camtrapdp_latest)
   )
 })